requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
jupyter
openpyxl
